requirements.txt
langchain
langchain_community
langchain_openai
python-dotenv
openai
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
groq
cassio
beautifulsoup4
langchain-groq
langchainhub
chromadb
faiss-cpu
wikipedia
arxiv
PyPDF2
langchain_google_genai

